fix(critical): recall() BFloat16序列化崩溃 - 记忆召回完全失效的根因
根因分析:
- recall() 在第243行做 query_for_hash.numpy().tobytes() 时，
  query_features 是 BFloat16 类型，CPU 不支持 BFloat16→numpy
- 导致 TypeError: Got unsupported ScalarType BFloat16
- 此错误被上层 try/except 捕获后静默忽略，
  导致 memory_context 始终为空，记忆信息无法注入到prompt

修复:
- hippocampus_system.py: query hash 计算前加 .float()
- hippocampus_system.py: dg/key/value features 序列化加 .float()
- ca3_memory.py: to_dict() 中所有 tensor 序列化加 .float()

这是记忆系统 0/6 召回率的最根本原因！

+++ hippocampus/ca3_memory.py
@@ -72,19 +72,19 @@ class EpisodicMemory:
             'consecutive_misses': self.consecutive_misses,
         }
         
-        # 安全序列化tensor：detach并转为numpy
+        # 安全序列化tensor：detach, 转float32, 再转numpy（避免BFloat16）
         if self.dg_features is not None:
-            result['dg_features'] = self.dg_features.detach().cpu().numpy().tolist()
+            result['dg_features'] = self.dg_features.detach().cpu().float().numpy().tolist()
         
         if self.key_features is not None:
-            result['key_features'] = self.key_features.detach().cpu().numpy().tolist()
+            result['key_features'] = self.key_features.detach().cpu().float().numpy().tolist()
         
         if self.value_features is not None:
-            result['value_features'] = self.value_features.detach().cpu().numpy().tolist()
+            result['value_features'] = self.value_features.detach().cpu().float().numpy().tolist()
         
         # 序列化语义 embedding
         if self.semantic_embedding is not None:
-            result['semantic_embedding'] = self.semantic_embedding.detach().cpu().numpy().tolist()
+            result['semantic_embedding'] = self.semantic_embedding.detach().cpu().float().numpy().tolist()
         
         return result
